requirements.txt
annotated-types==0.6.0
anyio==4.3.0
dnspython==2.6.1
fastapi==0.110.0
idna==3.6
motor==3.3.2
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.2
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
